Split lib/discovery.js out of lib/protocol.js (#80)
lib/protocol.js had grown to 999 lines mixing two distinct concerns:
the per-handler ProtocolHandler session (UDP/TCP connect, sendCommand,
runtime/info reads) and one-shot LAN discovery (UDP broadcast scan +
IPv4 helpers + family/model detection). They share `lib/modbus.js`
but are otherwise unrelated.

Move to lib/discovery.js:
- ipv4ToInt, isLocalSubnet, isPrivateBroadcast (IPv4 helpers from #61/#76)
- FAMILY_TAGS, detectInverterFamily, extractSerialNumber,
  extractModelName, parseDiscoveryResponse (from #57)
- discoverInverters (the function itself)

Stays in lib/protocol.js:
- ProtocolHandler class + private methods
- resolveCommAddr
- parseDeviceInfo (AA55 device-info reply parser — used by
  ProtocolHandler.readDeviceInfo)

lib/protocol.js re-exports the discovery functions for back-compat —
no consumer changes are *required*, only optional. nodes/discover.js
and test/discover-node.test.js are updated to import directly from
lib/discovery.js since that's the relevant module for them.

Net diff: protocol.js 999 → ~665 lines. New discovery.js ~310 lines.
No new tests (refactor only); existing 338 tests pass via re-exports.

Closes #80.
