+++ lib/protocol.js
@@ -7,7 +7,6 @@
 
 const dgram = require("dgram");
 const net = require("net");
-const os = require("os");
 const EventEmitter = require("events");
 
 // When no expectedLength is supplied, TCP `_sendCommandImpl` resolves after
@@ -18,82 +17,8 @@ const IDLE_BYTES_TIMEOUT_MS = 100;
 const { getFamilyConfig, parseSensorData } = require("./sensors");
 const modbus = require("./modbus");
 const { enhanceError } = require("./errors");
-
-/**
- * Convert a dotted-quad IPv4 string to a 32-bit unsigned integer (big-endian).
- * Returns null for input that isn't a valid IPv4 literal.
- * @param {string} ip
- * @returns {number|null}
- */
-function ipv4ToInt(ip) {
-    if (typeof ip !== "string") return null;
-    const parts = ip.split(".");
-    if (parts.length !== 4) return null;
-    let n = 0;
-    for (const p of parts) {
-        const v = Number(p);
-        if (!Number.isInteger(v) || v < 0 || v > 255 || /\D/.test(p)) return null;
-        n = (n * 256) + v;
-    }
-    // Force unsigned 32-bit.
-    return n >>> 0;
-}
-
-/**
- * Return true when `ipAddress` shares a subnet with any non-internal IPv4
- * interface on this host, or is loopback. Used to filter spoofed discovery
- * responses arriving from outside the local L2 segment.
- *
- * Defaults to false on parse errors so callers fail closed.
- * @param {string} ipAddress
- * @returns {boolean}
- */
-function isLocalSubnet(ipAddress) {
-    const peer = ipv4ToInt(ipAddress);
-    if (peer === null) return false;
-    // Loopback /8 — always trust.
-    if ((peer & 0xFF000000) >>> 0 === 0x7F000000) return true;
-    const ifaces = os.networkInterfaces();
-    for (const list of Object.values(ifaces)) {
-        if (!list) continue;
-        for (const iface of list) {
-            if (iface.family !== "IPv4" || iface.internal) continue;
-            const local = ipv4ToInt(iface.address);
-            const mask = ipv4ToInt(iface.netmask);
-            if (local === null || mask === null) continue;
-            if (((local & mask) >>> 0) === ((peer & mask) >>> 0)) return true;
-        }
-    }
-    return false;
-}
-
-/**
- * Return true for IPv4 broadcast/multicast/unicast addresses that are safe
- * to use as a discovery broadcast target without leaking the AA55 probe over
- * routed links to public hosts (#76):
- *   - 255.255.255.255 (limited broadcast — never leaves the local segment)
- *   - RFC1918 private ranges: 10/8, 172.16/12, 192.168/16
- *   - 169.254/16 link-local
- *   - 224/4 link-local multicast (224.0.0.0/24 specifically)
- * Any other address is treated as public and rejected.
- *
- * Returns false on parse errors so callers fail closed.
- * @param {string} ipAddress
- * @returns {boolean}
- */
-function isPrivateBroadcast(ipAddress) {
-    const addr = ipv4ToInt(ipAddress);
-    if (addr === null) return false;
-    if (addr === 0xFFFFFFFF) return true; // 255.255.255.255 limited broadcast
-    const top = (addr & 0xFF000000) >>> 24;
-    const second = (addr & 0x00FF0000) >>> 16;
-    if (top === 10) return true;                                // 10.0.0.0/8
-    if (top === 172 && second >= 16 && second <= 31) return true; // 172.16.0.0/12
-    if (top === 192 && second === 168) return true;             // 192.168.0.0/16
-    if (top === 169 && second === 254) return true;             // 169.254.0.0/16 link-local
-    if (top === 224 && second === 0) return true;               // 224.0.0.0/24 link-local multicast
-    return false;
-}
+// Discovery helpers moved to ./discovery in #80. Re-exported below for back-compat.
+const discovery = require("./discovery");
 
 /**
  * Resolve a user-supplied `commAddr` config value to a Modbus unit ID byte.
@@ -723,277 +648,18 @@ function parseDeviceInfo(payload) {
     return info;
 }
 
-/**
- * Discover GoodWe inverters on the network
- * @param {Object} options - Discovery options
- * @param {number} options.timeout - Discovery timeout in ms (default: 5000)
- * @param {string} options.broadcastAddress - Broadcast address (default:
- *   255.255.255.255). Rejected if it isn't an RFC1918/link-local/limited-
- *   broadcast address — see `isPrivateBroadcast`. Override the rejection by
- *   setting `options.allowPublicBroadcast: true` (you are then responsible
- *   for ensuring the AA55 probe does not leak over routed links).
- * @param {boolean} options.allowPublicBroadcast - Opt-out of the
- *   broadcastAddress private-range check (#76). Default false.
- * @param {boolean} options.acceptAnySource - When true, skip the local-subnet
- *   source-IP filter and accept discovery responses from any host. Default
- *   false. Only enable for routed setups where the inverter is reachable but
- *   not on the same L2 segment as the Node-RED host (you are then responsible
- *   for the trust boundary).
- * @returns {Promise<Array>} Array of discovered inverters
- */
-function discoverInverters(options = {}) {
-    return new Promise((resolve, reject) => {
-        const timeout = options.timeout || 5000;
-        const broadcastAddress = options.broadcastAddress || "255.255.255.255";
-        const acceptAnySource = options.acceptAnySource === true;
-        const allowPublicBroadcast = options.allowPublicBroadcast === true;
-
-        // Refuse to broadcast to public IPs unless the caller explicitly
-        // opted out (#76). The AA55 discovery probe contains nothing
-        // sensitive itself, but blasting it over routed links to the
-        // internet at large is a hygiene issue and can trigger firewall
-        // alerts for whoever owns the destination.
-        if (!allowPublicBroadcast && !isPrivateBroadcast(broadcastAddress)) {
-            const err = new Error(
-                "Refusing to broadcast discovery probe to non-private address " +
-                `${JSON.stringify(broadcastAddress)}. ` +
-                "Use a private (RFC1918), link-local, or limited-broadcast address, " +
-                "or set options.allowPublicBroadcast=true to override."
-            );
-            err.code = "INVALID_CONFIG";
-            reject(err);
-            return;
-        }
-
-        const discoveredInverters = [];
-        // Diagnostic counters (#78). Attached as a non-enumerable property
-        // on the resolved array so callers that ignore them still see a
-        // plain Array shape.
-        const diagnostics = {
-            nonLocalSubnet: 0,
-            invalidFrame: 0,
-            parseFailures: 0
-        };
-
-        const socket = dgram.createSocket("udp4");
-
-        let cleanedUp = false;
-        const cleanup = () => {
-            if (cleanedUp) return;
-            cleanedUp = true;
-            socket.removeAllListeners();
-            // Node 24+'s dgram throws "Not running" when close() is called on
-            // an unbound socket — which happens if send() errors before bind()
-            // completes (e.g. broadcastAddress fails DNS resolution).
-            try {
-                socket.close();
-            } catch (_err) {
-                // Already closed or never bound — nothing to release.
-            }
-        };
-
-        const resolveWithDiagnostics = () => {
-            Object.defineProperty(discoveredInverters, "diagnostics", {
-                value: diagnostics,
-                enumerable: false,
-                configurable: false,
-                writable: false
-            });
-            resolve(discoveredInverters);
-        };
-
-        const timeoutId = setTimeout(() => {
-            cleanup();
-            resolveWithDiagnostics();
-        }, timeout);
-
-        socket.on("error", (err) => {
-            clearTimeout(timeoutId);
-            cleanup();
-            reject(err);
-        });
-
-        socket.on("message", (msg, rinfo) => {
-            try {
-                // (1) Trust boundary: reject sources outside our local subnet
-                // unless the caller explicitly opted out. AA55 has no auth, so
-                // any LAN host that knows the protocol can advertise itself.
-                if (!acceptAnySource && !isLocalSubnet(rinfo.address)) {
-                    diagnostics.nonLocalSubnet++;
-                    return;
-                }
-                // (2) Validate AA55 framing (magic bytes + length + 16-bit
-                // checksum) before parsing. validateAA55Response would also
-                // check an expected response type, but discovery uses a fixed
-                // command so we only need structural validity here.
-                const validation = modbus.validateAA55Response(msg);
-                if (!validation.valid) {
-                    diagnostics.invalidFrame++;
-                    return;
-                }
-                const inverter = parseDiscoveryResponse(msg, rinfo.address);
-                if (inverter) {
-                    const exists = discoveredInverters.some(inv => inv.ip === inverter.ip);
-                    if (!exists) {
-                        discoveredInverters.push(inverter);
-                    }
-                }
-            } catch (err) {
-                // Continue listening but track the failure for diagnostics (#78).
-                diagnostics.parseFailures++;
-            }
-        });
-
-        // Use explicit listening/error events instead of bind(callback).
-        // bind() only invokes its callback on success — if the socket can't
-        // bind (e.g. port already in use during a second discovery), the
-        // callback never fires and we'd time out with empty results. With
-        // the on("error") handler above, bind failures already reject; this
-        // makes the intent explicit (#78).
-        socket.once("listening", () => {
-            socket.setBroadcast(true);
-            socket.send(
-                modbus.AA55_COMMANDS.DISCOVERY,
-                8899,
-                broadcastAddress,
-                (err) => {
-                    if (err) {
-                        clearTimeout(timeoutId);
-                        cleanup();
-                        reject(err);
-                    }
-                }
-            );
-        });
-        socket.bind();
-    });
-}
-
-/**
- * Inverter family classification by model-name substring (#57). Ported from
- * upstream marcelblijleven/goodwe `goodwe/model.py` model tag tables. Tags
- * are matched as substrings (case-insensitive) anywhere in the model name —
- * GoodWe model strings vary in format ("GW5000-ETU", "GW10K-DT", "GW25KMT")
- * so a strict prefix match would miss variants.
- *
- * Order matters where families share characters; the iteration below tries
- * DT first (because some DT tags like "DTU" overlap with "ET..." substrings
- * if matched loosely), then ES, then ET (which catches the platform-745
- * hybrids).
- */
-const FAMILY_TAGS = {
-    DT: ["DTU", "DTS", "D-NS", "DNS", "KMT", "XSU", "XS-", "MS-", "MSU"],
-    ES: ["ESU", "EMU", "BPU"],
-    ET: [
-        "ETU", "EHU", "BTU", "BHU",   // hybrid ET/EH/BT/BH
-        "ESA", "EHA",                 // ESA/EHA-G20 (platform 745 NAH)
-        "ARB", "URB", "EBR",          // platform 745
-        "NAH", "HMB", "HBB", "SPN"    // platform 745 hybrids
-    ]
-};
-
-/**
- * Parse discovery response (AA55 device-info reply broadcast by inverters)
- * @param {Buffer} data - Full AA55 frame as received
- * @param {string} ipAddress - IP address of the responder
- * @returns {Object|null} Parsed inverter info, or null if the frame is not
- *   a recognizable AA55 device-info reply.
- * @private
- */
-function parseDiscoveryResponse(data, ipAddress) {
-    try {
-        // Validate framing again at the parser boundary (the caller in
-        // discoverInverters also validates — this is defense in depth so
-        // direct callers cannot skip the check).
-        const validation = modbus.validateAA55Response(data);
-        if (!validation.valid) return null;
-
-        const payload = modbus.extractAA55Payload(data);
-        // Need at least model name (10) + serial (16) = 26 bytes.
-        if (payload.length < 26) return null;
-
-        const modelName = payload.slice(0, 10).toString("ascii").replace(/\0/g, "").trim();
-        const serialNumber = payload.slice(10, 26).toString("ascii").replace(/\0/g, "").trim() || "UNKNOWN";
-        const family = detectInverterFamily(modelName);
-
-        return {
-            ip: ipAddress,
-            port: 8899,
-            family,
-            serialNumber,
-            modelName
-        };
-    } catch (err) {
-        return null;
-    }
-}
-
-/**
- * Detect inverter family by matching the model name against the FAMILY_TAGS
- * table. Returns null when no tag matches — the discover node falls back to
- * a default ("ET") but null preserves the "unknown" signal in the payload.
- *
- * @param {string} modelName - Model name string (e.g. "GW5000-ETU")
- * @returns {string|null} "ET" | "DT" | "ES" | null
- * @private
- */
-function detectInverterFamily(modelName) {
-    if (!modelName || typeof modelName !== "string") return null;
-    const upper = modelName.toUpperCase();
-    for (const family of Object.keys(FAMILY_TAGS)) {
-        if (FAMILY_TAGS[family].some(tag => upper.includes(tag))) return family;
-    }
-    return null;
-}
-
-/**
- * Extract the serial number from an AA55 device-info response payload.
- * @param {Buffer} data - Full AA55 frame
- * @returns {string} 16-byte serial number trimmed of nulls/whitespace,
- *   or "UNKNOWN" if the frame is too short / invalid.
- * @private
- */
-function extractSerialNumber(data) {
-    try {
-        const validation = modbus.validateAA55Response(data);
-        if (!validation.valid) return "UNKNOWN";
-        const payload = modbus.extractAA55Payload(data);
-        if (payload.length < 26) return "UNKNOWN";
-        return payload.slice(10, 26).toString("ascii").replace(/\0/g, "").trim() || "UNKNOWN";
-    } catch (err) {
-        return "UNKNOWN";
-    }
-}
-
-/**
- * Extract the model name from an AA55 device-info response payload.
- * @param {Buffer} data - Full AA55 frame
- * @returns {string} 10-byte model name trimmed of nulls/whitespace,
- *   or empty string if the frame is too short / invalid.
- * @private
- */
-function extractModelName(data) {
-    try {
-        const validation = modbus.validateAA55Response(data);
-        if (!validation.valid) return "";
-        const payload = modbus.extractAA55Payload(data);
-        if (payload.length < 10) return "";
-        return payload.slice(0, 10).toString("ascii").replace(/\0/g, "").trim();
-    } catch (err) {
-        return "";
-    }
-}
 
 module.exports = {
     ProtocolHandler,
-    discoverInverters,
     parseDeviceInfo,
-    // Exported for unit tests; not part of the public Node-RED API.
-    isLocalSubnet,
-    isPrivateBroadcast,
-    ipv4ToInt,
-    detectInverterFamily,
-    extractModelName,
-    extractSerialNumber,
-    parseDiscoveryResponse,
+    // Discovery functions live in ./discovery (#80); re-exported here so
+    // existing consumers continue to work without changing their import.
+    discoverInverters: discovery.discoverInverters,
+    isLocalSubnet: discovery.isLocalSubnet,
+    isPrivateBroadcast: discovery.isPrivateBroadcast,
+    ipv4ToInt: discovery.ipv4ToInt,
+    detectInverterFamily: discovery.detectInverterFamily,
+    extractModelName: discovery.extractModelName,
+    extractSerialNumber: discovery.extractSerialNumber,
+    parseDiscoveryResponse: discovery.parseDiscoveryResponse
 };
